Replace the uneven-streams trace
The previous trace did have streams of different ranges, but all the
actual events were part of the intersection of all streams, so it was
not very useful for testing it event-wise.

Replace it with another trace that has a very different event count
between all streams and the intersection range (for reference, this
trace was taken by hot-plugging a CPU while tracing).

Version 1.7.1
("breaking" change, but nothing committed depended on the previous trace
so it should be safe to just do a minor bump).

Signed-off-by: Alexandre Montplaisir <alexmonthy@efficios.com>