Add invalid trace files for ftrace security testing
This commit adds 2 new binary ftrace files. In most cases, each section
in the header of binary ftrace files contains a section header, and a
section size. The size of the section allows the parser to read and
process the section data.

The 2 added ftrace files are injected with bad data, specifically bad
section sizes for the header page section and the first cpu section.
They are used for security tests and prevention of injection attacks. If
the parser uses bad section sizes to read section data, it might go out
of the file size limit and get into unauthorized memory locations.

Add the missing blank javadoc lines in FtraceTestTrace, so the default
Eclipse formatter does not do so while adding extra trailing space
characters to these (wrongfully).

Change-Id: Ia557b83416de83d0751cd33fef031bf4db09c0ef
Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/tracecompass-test-traces/+/190075
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>