Rename binary ftrace test trace recorded on a multiple-core machine
This commit renames the valid binary ftrace test trace that was recorded
on a multiple-core machine. Previously, the trace name was
TEST_2_6_4_CPU, which was confusing for readers since it can be
understood as FTrace version 2.6.4. However, the FTrace version was 2.6
and the number of CPUs was 4. It is now renamed to
TEST_2_6_MULTIPLE_CPUS, which reflects clearly that it was recorded on a
machine with multiple cores.

Change-Id: I228512fa4d1d87d34b8551f374ad560b038b741e
Signed-off-by: Hoang Thuan Pham <hoang.pham@calian.ca>
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/tracecompass-test-traces/+/190449
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Marco Miller <marco.miller@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>