Add an OTF2 trace with double counters
Add an OTF2 trace with ru_stime and ru_utime counters. These counters
are stored as integers (number of microseconds) in the trace. However,
they should be stored as double in the state system to be displayed with
the generic counters analysis since the base unit is a second. This can
be used to test the support for double metrics. The tracer is Score-P
7.1.

This change depends on [1,2]:

[1]https://git.eclipse.org/r/c/tracecompass.incubator/org.eclipse.tracecompass.incubator/+/193333
[2]https://git.eclipse.org/r/c/tracecompass/org.eclipse.tracecompass/+/193595

Signed-off-by: yoann-heitz <yoann.heitz@polymtl.ca>
Change-Id: Iaa9fcadff6de08acfb10f758fcc913cc739ccb2b
Reviewed-on: https://git.eclipse.org/r/c/tracecompass/tracecompass-test-traces/+/193594
Tested-by: Trace Compass Bot <tracecompass-bot@eclipse.org>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>